requirements.txt
arrow==1.3.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
botcity-framework-base==0.4.4
botcity-framework-core==1.2.1
botcity-framework-web==0.9.0
botcity-maestro-sdk==0.5.1
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.6
contourpy==1.2.1
cookiecutter==2.6.0
cycler==0.12.1
distlib==0.3.8
distro==1.9.0
filelock==3.15.4
fonttools==4.53.1
idna==3.7
importlib_metadata==8.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
msedge-selenium-tools==3.141.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
panda==0.3.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
psutil==6.0.0
Pygments==2.18.0
pynput==1.7.7
pyparsing==3.1.2
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-xlib==0.33
pytz==2024.1
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
selenium==3.141.0
setuptools==72.2.0
six==1.16.0
soupsieve==2.6
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==1.26.19
virtualenv==20.26.3
webdriver-manager==4.0.2
zipp==3.20.0
